requirements.txt
anthropic>=0.50.0
chromadb>=0.5.5
python-dotenv>=1.0.1
